requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
pytz==2023.3
psycopg2-binary==2.9.7
